I created the row appbare of details_screen
