+++ lib/features/home/presentaion/screens/widgets/book_details_screen_appbare_widget.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class BookDetailsScreenAppbareWidget extends StatelessWidget {
+  const BookDetailsScreenAppbareWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: [
+        IconButton(onPressed: () {}, icon: const Icon(Icons.close)),
+        IconButton(
+            onPressed: () {}, icon: const Icon(Icons.shopping_cart_outlined))
+      ],
+    );
+  }
+}
